ReactPdfJs now accepts a `containerRef` prop to scale against
The way this is achieved is by comparing the pdf page's ratio
(width/height) to the containers ratio.

If the container has a larger ratio we base our scale value on the
dividend of the containers width and canvas's width.

We do the opposite (and default) for when the container has al ower
ratio than the canvas.